display the aggregations as integers

+++ Data_Visualizatioin_Matplotlib.py
@@ -63,11 +63,11 @@ print(auto_new.dtypes)
 print()
 
 # group by <Body Style>, aggregate(max, min, avg)
-print(auto_new.groupby('Body Style').Price.agg(['min','max','mean']) )
+print(auto_new.groupby('Body Style').Price.agg(['min','max','mean']).astype(int) )
 
 # group by <Make>, aggregate(avg price)
 print()
-print(auto_new.groupby('Make').Price.agg(['mean']) )
+print(auto_new.groupby('Make').Price.agg(['mean']).astype(int))
